LPS-159150 Rename Icons
