+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/images/TestrayLogo.tsx
