+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/images/index.ts
@@ -1,0 +1,18 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+export {default as ForwardIcon} from './ForwardIcon';
+export {default as TestrayIcon} from './TestrayIcon';
+export {default as TestrayIconBrand} from './TestrayIconBrand';
+export {default as TestrayLogo} from './TestrayLogo';
